Refactor para centralizar metodos comuns
