--- src/components/InclusaoDeAlimentacao/DRE/PainelPedidos/helper.js
@@ -1,7 +1,0 @@
-export const formatarPedidos = pedidos => {
-  return pedidos.map(pedido => {
-    let novoPedido = pedido;
-    novoPedido["checked"] = false;
-    return novoPedido;
-  });
-};
