+++ src/components/SuspensaoDeAlimentacao/components/CardHistorico/index.jsx
@@ -3,7 +3,7 @@ import { connect } from "react-redux";
 import { Field, reduxForm, formValueSelector } from "redux-form";
 import { Collapse } from "react-collapse";
 import { Redirect } from "react-router-dom";
-import { obtemDataSolicitacao } from "../../Terceirizada/PainelPedidos/helper";
+import { obtemDataSolicitacao } from "../../../../helpers/painelPedidos";
 import "./style.scss";
 import {
   TERCEIRIZADA,
